% This example requires vetools at least version 1.3.29
\name{estados.venezuela}
\alias{estados.venezuela}
\alias{Abb}
\alias{Name}
\alias{Name.UTF8}
\alias{shape.name}
\title{Data.frame of the states that conform Venezuela}
\description{
        Shows the names and abbreviations of the states of venezuela.
}
\author{
	A.M. Sajo-Castelli
}
\keyword{ estados }
\keyword{ venezuela }
