\name{xts2ts}
\alias{xts2ts}

\title{
Time-serie convertion routine
}
\description{
Converts from class \code{"xts"} to class \code{"ts"}
}

\usage{
xts2ts(b.xts)
}
\arguments{
  \item{b.xts}{ object of class \code{"xts"} to convert to class \code{"ts"}}
}

\value{
returns an object of class \code{"ts"}
}

\author{
A.M. Sajo-Castelli
}

\seealso{
\pkg{\link{vetools}},
\code{\link{diffmonths}},
\code{\link{tssum}},
\code{\link{diasdelmes}},
\code{\link{time2ym}},
\code{\link{ym2time}},
\code{\link{m12}}.
}

\keyword{ vetools }
\keyword{ time }
\keyword{ tssum }
